+++ src/components/talks/components/SeriesList/index.tsx
@@ -1,23 +1,23 @@
 import React, { PureComponent } from "react"
 
-import type { Series } from "../../types"
+import type { Series, Sermon } from "../../types"
 import Grid from "../Grid"
 import SeriesListItem from "./SeriesListItem"
 
 interface Props {
     serieses: Array<Series>
-    onSelectSeries: (seriesId: string) => void
+    shouldHighlightSermon: (sermon: Sermon) => boolean
 }
 
 class SeriesList extends PureComponent<Props> {
     render() {
-        const { serieses, onSelectSeries } = this.props
+        const { serieses, shouldHighlightSermon } = this.props
         return (
             <Grid
                 items={serieses}
                 keyExtractor={series => series.id}
                 renderItem={series => (
-                    <SeriesListItem onClick={onSelectSeries} series={series} />
+                    <SeriesListItem series={series} shouldHighlightSermon={shouldHighlightSermon}/>
                 )}
             />
         )
